♻️ refactor(move): moved AbbreviatedAvatar into sidebar folder
