+++ frontend/components/(dashboard)/sidebar/AbbreviatedAvatar.tsx
